Adds a function to cutflow.py to print multiple cutflows as table columns
New function in class cutflow (print_multi_table) to print multiple cutflows at once. This was tested and worked with in bgCutflow and run at the end of backgroundAnalysis. Many more samples were included in backgroundAnalysis as well.
The bugged histogram "lj_pfiso" was uncommented, since the bug was resolved.