Finalizing commit for pull
The main files are bgCutflow (where some cutflows are shown and tested), bgAnalysis, where histogram-generating functions are written and compared to background, and bg_FileTests, which was mainly used to try to find the errors in some background files. A new function was also added to cutflow.py to put multiple cutflows on the same table.