Bg monte carlo (#32)
* Update add_ntuples to handle FFV2 directories and add FFV2 ntuples to ntuple_locations

* Update ffschema to work with v2 and v4 ntuples

* Update test_FFSchema notebook

* Warn user and skip histogram if calling fill function produces AttributeError

* Minor updates to hists, cuts, and objects to allow use of v2 ntuples

Need to check if any updates alter v4 results

* Fix bug preventing explicit pT ordering

* Verify that results with V4 ntuples are unchanged

* add optional argument to limit the number of files to run over from each dataset

* Fix handling of multiword object branches to work with v2 ntuples (signal and bg)

* Add protections to warn user and skip hists and cuts that use non-existant objects and attributes

* Added background sample sets to notebooks, ran some samples to find cutflow

* Background analysis, still buggy

This commit is mostly in the bgCutflow and background_analysis, and adds a "test" histogram collection. These are still errors in running the ljpfiso histogram and in a few QCD samples, but with these commented out, it still works.

* Bug Fixes for bgSamples

Adds a file to test different samples, resolves issues with lj_pfiso histogram and some QCD samples. Dask still has an error, since it gets code from GitHub instead of local.

* Changed base for Dask workers

Changes the file scaleout.py to be based on the bgMonteCarlo branch

* Updated background_analysis, reverted scaleout.py change

Dask has an error currently, so this commit reverts an earlier change to scaleout.py in hopes of fixing it. (Also gives a more comprehensive set of histograms in bg analysis)

* Adds a function to cutflow.py to print multiple cutflows as table columns

New function in class cutflow (print_multi_table) to print multiple cutflows at once. This was tested and worked with in bgCutflow and run at the end of backgroundAnalysis. Many more samples were included in backgroundAnalysis as well.
The bugged histogram "lj_pfiso" was uncommented, since the bug was resolved.

* Finalizing commit for pull

The main files are bgCutflow (where some cutflows are shown and tested), bgAnalysis, where histogram-generating functions are written and compared to background, and bg_FileTests, which was mainly used to try to find the errors in some background files. A new function was also added to cutflow.py to put multiple cutflows on the same table.

* Rework of postprocessor

Changes the postprocessor in analysis and cutflow notes such that copied dict objects no longer reference the original.

* Adds bgAnalysis and bg_FileTests to push

* Added background sample sets to notebooks, ran some samples to find cutflow

* resolved some conflicts

* resolved conflict in hist_collections

* Changed base for Dask workers

Changes the file scaleout.py to be based on the bgMonteCarlo branch

* Updated background_analysis, reverted scaleout.py change

Dask has an error currently, so this commit reverts an earlier change to scaleout.py in hopes of fixing it. (Also gives a more comprehensive set of histograms in bg analysis)

* Adds a function to cutflow.py to print multiple cutflows as table columns

New function in class cutflow (print_multi_table) to print multiple cutflows at once. This was tested and worked with in bgCutflow and run at the end of backgroundAnalysis. Many more samples were included in backgroundAnalysis as well.
The bugged histogram "lj_pfiso" was uncommented, since the bug was resolved.

* Finalizing commit for pull

The main files are bgCutflow (where some cutflows are shown and tested), bgAnalysis, where histogram-generating functions are written and compared to background, and bg_FileTests, which was mainly used to try to find the errors in some background files. A new function was also added to cutflow.py to put multiple cutflows on the same table.

* Rework of postprocessor

Changes the postprocessor in analysis and cutflow notes such that copied dict objects no longer reference the original.

* Adds bgAnalysis and bg_FileTests to push

* Compares background sources

Adds the files bgSampleComparison to draw comparisons between individual sources of background signals. Also adds a function cut_breakdown() to the cutflow.py in order to get individual cut paths for plotting. cuts.py and objects.py had changes to attributes present in v4 ntuples, this commit changes them back. It should be obsolete with introduction of LLPNanoAODs anyway.

* Delete sidm/studies/Background Sample List (Reference).txt

Removes this file, it was kept for reference, and data inside is held in ntuple locations

* Removes processor.coffea file

Not sure what this file is, it doesn't seem to have any data.

* Removes whitespace in histogram.py

* Adds blank line to processor

* Removed commented line

* Deletes processor.coffea

Not sure what this file is or what it's doing here.

* Removed background sample list.txt

* Brought cuts and objects back in line with origin

* Reverted changes to outputs of multiple_samples and scale_out.ipynb

* Removed test histogram collection

* Reverts changes from running notebooks

* Brings leptonKinematics up to upstream version

* Reverts changes from running for sample_management and test_FFschema

* Gets upstream version of test_FFschema to revert running changes

* Runs sample_management to match output

* Overwriting sample_management with upstream version

---------

Co-authored-by: Bryan Cardwell <bryan.cardwell@cern.ch>
Co-authored-by: btcardwell <bryan.cardwell@gmail.com>